+++ consumer_open.go
@@ -48,6 +48,7 @@ func openConsumerWith(
 	if err := connection.Validate(); err != nil {
 		return nil, err
 	}
+	config = ownConsumerConfig(config)
 	if err := config.Validate(); err != nil {
 		return nil, err
 	}